fixed some issues with parent/child relationships being circular.
it is still possible to hack the page form data as a root admin and
get the servlet engine into a bad state as it seems that i was not
able to close all the holes as a result of what happens once a circular
dep is created, but it would be really difficult to do this so i'm not
going to spend a few days on trackign down the issue.

summary: the ui doesn't allow you to create a circular dep anymore.
         the action doesn't allow you to save it to the database.
         the worst that could happen is bad in memory data that can be
             cleared on a container restart.

-jon


git-svn-id: http://scarab.tigris.org/svn/scarab@4363 3c7165bb-d2ef-0310-b724-bafd9a86081f