added new screen (reintroduced discarded screen, anyway):: 2.6.1.1 Bulk edit issues
This screen allows users to select multiple issues from a query results, and edit their attributes as if they were one issue. As noted in the "alert" at the top of this screen, attribute values expressed as "---" on this screen are values that differ within the selected issues. altering the "---" changes the value for all issues being edited. Leaving a value as "---" does not change the value in any of the issues being edited.

Appropriate UI widget ("Bulk edit") already existed on the query results UI template (2.6.0)

Updated the index with this screen as well.

thanks
B


git-svn-id: http://scarab.tigris.org/svn/scarab@4630 3c7165bb-d2ef-0310-b724-bafd9a86081f