need to keep track of the current module and issuetype on a per thread basis as a user could have multiple windows open.
It is possible to consider each browser window a user has open as a separate
thread.  Even though they may actually use the same thread it will never be
at the same time.  So as long as a value is set for each new request, storing
that value in a ThreadLocal provides an abstraction that allows us to not
have to add the request key for every method in user that involves
state related to a particular client window.

This is useful for not introducing complexity into the business api
that is related to http.

For data that is not set for each request (session data used between requests)
we still must use Maps to keep data for each browser separate.  But this
can all be done with private scoped code, and the public api does not expose
the thread/request requirements.

I have used this new way with the CurrentMITList which is used while building
a cross module query.  It should be used in other places where state is being
stored in the user object.  But issues such as stale data and invalid keys due
to bookmarking need to be addressed.  These issues are present in the previous
system as well, so they need to be considered.


Issue number:
Obtained from:
Submitted by:
Reviewed by:

Issue number:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://scarab.tigris.org/svn/scarab@4950 3c7165bb-d2ef-0310-b724-bafd9a86081f