first cut at showing issue id's mentioned in comments as links.
Links of the usual form, like PACS101, as well as numbers prefixed with
issue#1, bug 1, or artifact #1.  In addition to these three hardcoded qualifiers
the issue type names can be used.  So that Defect#PACS1 would be a valid id.

The links must be rendered within the template to avoid the filtering, so the
results are passed to the template as a list of Strings and size=2 lists
which a macro assembles into the comment with links added.

Issue number:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://scarab.tigris.org/svn/scarab@5604 3c7165bb-d2ef-0310-b724-bafd9a86081f