Enable per-database-type configuration.
File default.${scarab.database.type}.properties is loaded if found.

In this way, properties that differ between databases can be automated.

Prime example (which motivated this) is setting the validation query differently for Oracle.

With this in place, it is not necessary to remember to place this into build.properties - simply setting scarab.database.type to oracle handles it.


git-svn-id: http://scarab.tigris.org/svn/scarab@5813 3c7165bb-d2ef-0310-b724-bafd9a86081f