a far better performing version of getSiteName()...
note that: I used a static variable to hold the reference so that
the configuration isn't read each and every time scarabG.SiteName
is referenced (no need since this is a configuration option
that is only needed once per startup instead of each time).
i also simplified the logic quite a bit by simply
checking for null instead of doing string comparisons.

FIXME: someone might want to i18n the output text so that it changes
depending on the browser. this should be a feature which is added in the
future and the method should be moved to scarabR instead in that case since
it is a request based feature instead of a global tool feature.

-jon


git-svn-id: http://scarab.tigris.org/svn/scarab@5885 3c7165bb-d2ef-0310-b724-bafd9a86081f