Submitted by:  Thierry Lach
Patches to get datadump, datadtd, and datasql executing.

Datadump does not appear to work correctly for Oracle, but this appears to be a torque issue (null column values seem to repeat the data from the previous column).

Datasql runs but with the data I was using encounters an out-of-memory abort.

But prior to this change, they didn't even execute.


git-svn-id: http://scarab.tigris.org/svn/scarab@5970 3c7165bb-d2ef-0310-b724-bafd9a86081f