previously the permission to perform an action was tied to the template that is going to be displayed.  This is not the best system as actions can be invoked via the url independent of the template.
This implements independent authorization of the user for the action.  It is
probable that actions will have more fine-grained permission checks
within the code, but this permission will need to be met prior to being able to
execute the action, and it is set via the property file.

It is currently set very lenient as only one action has the permission defined.
All actions that require login need to have a permission defined, then the
method should be switched to disallow any action that does not have the
permission defined, so that any that are missing are more easily detected.

Issue number:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://scarab.tigris.org/svn/scarab@5977 3c7165bb-d2ef-0310-b724-bafd9a86081f