From: Jean-Fran�ois El Fouly <jean-francois.elfouly@wanadoo.fr>
I suspect there is an erroneous carriage return in ScarabBundle_en_US.properties, line 1410 (after the key UQT16).

I also note in this help entry that the original version said that &amp;&amp; could be used instead of AND in a text query. In the new version, there is just one &.
Who's right ?

&amp;&amp;


git-svn-id: http://scarab.tigris.org/svn/scarab@6212 3c7165bb-d2ef-0310-b724-bafd9a86081f