improve the performance of search results by only performing the sql query once and caching the result for the rest of the request.
Also limiting the target issues of the form buttons to only those currently
being shown.  Then we do not have to list every issue id, which for a large
list can be very expensive.  It can also allow the use of limit and offset on
the sql query to further improve efficiency though that is not implemented yet.


git-svn-id: http://scarab.tigris.org/svn/scarab@6345 3c7165bb-d2ef-0310-b724-bafd9a86081f