pulling most of the data to display an issue list from a query in one sql statement. I also rewrote the part that was making a second select to get issues which did not have values for the sort attribute.  However doing so and also always putting the null valued attributes at the end precludes using limit/offset.  If that is added we will need to revert the attribute sort to use a second query or allow the null fields to be sorted first for one of the polarities.
These changes reduce the response time for a large resultset about 6-8X.  Not
quite as high as I expected, so there is still some debugging in as I am trying
to determine if there is some other resource hog, I have not found.  But it would be good to get other db's tested if anyone is game.


git-svn-id: http://scarab.tigris.org/svn/scarab@6438 3c7165bb-d2ef-0310-b724-bafd9a86081f