added a new method getActivitiesWithNullEndDate
added some caching

add the ability to hideDeleted with the getChildren/getParent dependencies

set some methods to private which should not have been public

don't pass in empty attachment objects...waste of memory and just doesn't work

fix up the doDeleteDependency functionality (renamed some variables,
throw ScarabException so that it can be caught specially)

make it so that we always delete an old dependency when we change
the dependency type and then create a new record to reflect the new
dep type. this will create more rows in the table, but will better
reflect the history of the issue.

make sure that there is data in the attachment before creating it...
no need to create empty attachments

deprecate the workflow method which doesn't even use the attachment
object which is passed into it. elicia, please check to make sure
that this doesn't screw any of your stuff up...


git-svn-id: http://scarab.tigris.org/svn/scarab@6725 3c7165bb-d2ef-0310-b724-bafd9a86081f