Issue number: pcn#14186
Modified Files:
	src/conf/cache.ccf
	src/java/org/tigris/scarab/om/IssueTypeManager.java
 	src/java/org/tigris/scarab/om/QueryManager.java
 	src/java/org/tigris/scarab/om/QueryPeer.java
 	src/java/org/tigris/scarab/om/ScarabModule.java
----------------------------------------------------------------------
Adding more jcs caching.

cache.ccf - properties for the IssueType and Query caches.
IssueTypeManager.java - enable cache
QueryManager.java - enable cache and invalidate all if one query changes
the coarse invalidation limits the utility of this cache, but it is likely
to still save a number of db selects.
QueryPeer.java - use the cache on the method that is called to populate
the query nav element that is on every page. Also added some string constants
and renamed 'me' to 'module'.
ScarabModule.java - torque bug skips the fourth element of the key, so
make it null.


git-svn-id: http://scarab.tigris.org/svn/scarab@6871 3c7165bb-d2ef-0310-b724-bafd9a86081f