Issue number: 14596 Modified Files:     src/java/org/tigris/scarab/actions/Search.java     src/java/org/tigris/scarab/om/QueryManager.java     src/java/org/tigris/scarab/om/QueryPeer.java ---------------------------------------------------------------------- * QueryPeer.java - skip array instantiation as the number of arguments   are covered by the overloaded get methods * QueryManager.java - don't need the CacheListener implementation * Search.java - checking that the pk is not the same when looking for   a duplicate query name.  it is okay to alter a query and keep the same   name.
Search.java also includes a number of performance altering changes.  Or at
least are idioms that I think should be used.

1.  Don't use array.size() > 0 when !array.isEmpty() will do.  If the size
is used in another statement and a local variable is created then arraySize > 0
might be appropriate.  Since we generally use ArrayList, the difference is
probably not great, but if the list implementation changed, size() could be
considerably more expensive than isEmpty().

2. Use an iterator when looping over a List.  Again with ArrayList the
difference may not matter, but it affects our ability to modify the underlying
implementation more difficult.  We have to analyze the usage later to
determine if the random access implied by get(i) is important.

3. If a method is going to give the same result regardless of position in the
loop, assign the value to a local variable outside the loop.
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://scarab.tigris.org/svn/scarab@7189 3c7165bb-d2ef-0310-b724-bafd9a86081f