Issue# 14591 Modified Files:     src/templates/email/ModifyIssue.vm Added Files:     src/templates/email/ModifyIssueSubject.vm ---------------------------------------------------------------------- * missed a subject template in previous commit and a fix for the user who modified the issue.
* Created a template for each of the email subjects that used to be
generated in the java code.  This is to make it easier to update the
contents of emails without having to recompile.  It might have been
cleaner to be able to generate the subject within the same email body
template that uses it, but some body templates are used with multiple
subjects.  Other than that the fulcrum service as currently written
does not make it possible.

* Created a velocity context to use for emails.  I did this to help improve
the distinction between the context meant for displaying the web response
and that used in emails.  Still passing the web related localization and
link tools, so the distinction is still blurred, but its a small improvement.


git-svn-id: http://scarab.tigris.org/svn/scarab@7227 3c7165bb-d2ef-0310-b724-bafd9a86081f