Issue number:  15742 Reviewed by:   John/DLR
org/tigris/scarab/actions/ReportIssue.java:

Limit the number of results to 25 because there is no paging for duplicates on wizard2 and there is no need to have more than that returned in most cases. This is defined as a constant so it is easily modified with a recompile.

Only run the dedupe query if there are attribute values set...

Use AttributeValue.isSet() to remove duplicated code.

org/tigris/scarab/util/word/IssueSearch.java:

Use AttributeValue.isSet() to remove duplicated code.


git-svn-id: http://scarab.tigris.org/svn/scarab@7389 3c7165bb-d2ef-0310-b724-bafd9a86081f