Issue number:  16763
built a new version of fulcrum from the .gz which is in the lib directory

this new version has changes to TemplateEmail which allow it to be overridden
easily to provide alternative implementations of its methods. the one to watch
is the handleRequest() method which allows one to specify a different engine
for rendering the body of the email.

i had to upgrade the commons-lang to latest cvs head because TemplateEmail
was referencing WordWrappers which was wrong...it has moved to
org.apache.commons.lang.WordWrapUtils


git-svn-id: http://scarab.tigris.org/svn/scarab@7714 3c7165bb-d2ef-0310-b724-bafd9a86081f