Submitted by:  Thierry Lach
Working with Scarab inside of an IDE is difficult if the source files are moved.
Setting "copy.peer.classes" inside of build.properties changes this behavior to a copy.
The default (with the property missing) is the traditional move.
This .cvsignore file keeps all of the regenerated files from being handled by CVS.


git-svn-id: http://scarab.tigris.org/svn/scarab@7919 3c7165bb-d2ef-0310-b724-bafd9a86081f