issue type link in ModuleQuery.vm changes current issue type leading to a couple issues Issue number: 19012 - if there are no active issue types, clicking the link will show links in the left nav that otherwise are not available. Issue number: 18365 - running a query after clicking the a link to an inactive issue type will run the query against an active issue type.
The link for a given issue type on ModuleQuery tab was meant to function as a
shortcut that should work the same as checking the checkbox (and only one
checkbox) next to the artifact and click Go button.  It should not have the side
effect of opening up navigation elements that would not otherwise be available.

Modified file:
* src/templates/screens/home/ModuleQuery.vm - give the link the same parameters
  that using the form would. Also shortened a boolean variable initialization.

Reviewed by: venkat


git-svn-id: http://scarab.tigris.org/svn/scarab@8110 3c7165bb-d2ef-0310-b724-bafd9a86081f