src/java/org/tigris/scarab/actions/admin/GlobalAttributeEdit.java private void mapAttribute(RunData data, TemplateContext context): This method is added in the class.It manages attribute to module/issue type mappings, when the user clicks on the "save"  button.Previously the attribute module/issuetype mapping was done only when the user clicks on the "done"button . This method also verifies whether the attribute mapping is already done or not. Based on the above , the data will be inserted into the database.
doCancel method: In the docancel method the contents which does the attribute
module/issue type were removed. Because it is inapproriate to do the mapping
when the user clicks on the cancel button.
********************************************************************************
src/java/org/tigris/scarab/om/AttributeGroup.java

public boolean hasAttribute(Attribute attribute): This method is added in the
class.It checks whether an attribute group has a particular attribute.This method
is called in the mapAttribute method of the above class.

********************************************************************************
src/java/org/tigris/scarab/om/Attribute.java

public boolean hasMapping(Module module, IssueType issueType): This method is
added in the class. It checks whether an attribute is already mapped to the
module and the issue type. It returns true if the attribute is already mapped.

public boolean hasMapping(IssueType issueType):This method is also added
in the class. It checks whether an attribute is already mapped to the issue
type. It returns true if the attribute is already mapped.

**********************************************************************************
src/templates/screens/admin/UserAttributeEdit.vm

Issuetypeid is passed as a hidden parameter,so that when the last template is displayed,
it shows all the details of the particular issuetype object.
************************************************************************************
 src/templates/screens/admin/GlobalAttributeEdit.vm

Attribute group Id is passed as a hidden parameter. so that when the last template is
displayed, it shows all the details of the particular attribute group.

Issue: PCN18803
Submitted by: irk@collab.net
Reviewed by: Elicia


git-svn-id: http://scarab.tigris.org/svn/scarab@8141 3c7165bb-d2ef-0310-b724-bafd9a86081f