*src/i18n/ScarabBundle_en_US.properties
IssueTypeUnavailable: This key is added. The value is an alert
message when the issue type is made inactive.

*src/java/org/tigris/scarab/actions/ReportIssue:

checkIssueTypeStatus: This method is added. This method checks
the current status of the issue type. Some time the admin may
make the issue type(IT1) inactive, when the user entering the issue
of type IT1. At that time this method gives an alert message to the
user that the issue type is not available.

This method is called whenever the user clicks on the next button or
submit button.

*src/templates/screens/entry/Wizard1.vm
*src/templates/screens/entry/Wizard2.vm
*src/templates/screens/entry/Wizard3.vm

In all the above classes the hidden parameter Current_ISSUE_TYPE
was removed, so that this issue type will not be appended with the
URL.


git-svn-id: http://scarab.tigris.org/svn/scarab@8244 3c7165bb-d2ef-0310-b724-bafd9a86081f