*src/java/org/tigris/scarab/om/Issue.java
#setAttributeValues(ActivitySet,HashMap,Attachment,ScarabUser,boolean)-removed.
#setAttributeValues(ActivitySet,HashMap,Attachment,ScarabUser)- Implementation in the overloaded
method brought back to this method and some indentation errors corrected.
Changed so that a call to "isTemplate()" is used in place of the boolean argument "isIssueTemplate".
This change in fact defeated the purpose of the overloaded method and hence it was removed.

*src/java/org/tigris/scarab/actions/TemplateList.java
#doEditvalues()-Both old and new values are trimmed before checking for
equality or length. Also the arguments to the call to the Issue.setAttributeValues()
corrected.
Suggested by: lacostej@coffeebreaks.org
Submitted by: venkat@collab.net
Original issue: PCN 20483


git-svn-id: http://scarab.tigris.org/svn/scarab@8531 3c7165bb-d2ef-0310-b724-bafd9a86081f