Fixes issue SCB1065:
  http://scarab.collab.net/scarab/issues/id/SCB1065

Whenever the admin changed their details, the user object in the
session was not being updated. Then when they logged out, this
user object was being saved into the database, wiping out the
previous changes.

So, whenever changes are made to the logged in user, the session
object is now updated as well. I have also removed the field for
changing a user's password as this caused problems whenever the
admin tried to change their password. When it comes down to it,
if a user forgets their password, they can go through the
"Forgot Password" page to get a new one. And the admin can change
their own password via the "Change Password" screen, which works
in a more consitent way than the "Edit User" screen.

Files modified:

  src/java/org/tigris/scarab/actions/admin/ManageUser.java
  src/webapp/WEB-INF/templates/screens/admin/EditUser.vm
  xdocs/changes.xml


git-svn-id: http://scarab.tigris.org/svn/scarab@9030 3c7165bb-d2ef-0310-b724-bafd9a86081f