These are a bunch of fixes related to Intake. First, the Fulcrum ValueParser is a different interface to Turbine's ValueParser, so ScarabRequestTool has to create an adapter between the two. Turbine itself had to be changed as well in order to support initialisation of IntakeTool via a ValueParser as well as a RunData object.
Second, the behaviour of Group.setProperties() changed such that
the destination object's properties were overwritten with defaults
whenever the associated Field was not set. This was causing all
sorts of havoc, so now I have added a Group.setPropertiesNoOverwrite()
method in the intake project. This uses the same semantics as provided
by the implementation in Turbine 3.

Finally, fields defined as multi-value "int" in the intake
configuration pass an int[] to the associated object property now,
rather than Integer[]. So I have had to change some setters to
accomodate this change.



git-svn-id: http://scarab.tigris.org/svn/scarab@9815 3c7165bb-d2ef-0310-b724-bafd9a86081f