A couple of fixes here. First, I have reactivated the UploadService which is actually required by Turbine to parse parameters passed in by a POST. If the upload service is not loaded, Scarab does very little as none of its actions are invoked!
Second, I have now removed all references to $renderer.render() in
the VM files as this feature is not available in Turbine 2.4. Any
occurrences have been replaced with macros which are basically copies
of the old templates, e.g. ViewIssueBody.vm has been moved into
ViewIssueMacros.vm as the viewIssueBody() macro. Unfortunately, this
has left us with an advancedQueryScreen() macro *and* the
AdvancedQuery.vm screen, but I think removing the screen will break
certain templates that do something like:

  $link.setPage("AdvancedQuery.vm")

So for now, we have to keep the screen template and the macro in sync
manually.


git-svn-id: http://scarab.tigris.org/svn/scarab@9817 3c7165bb-d2ef-0310-b724-bafd9a86081f