Two fixes here. First, the root user has been made initial owner of the Global module. This fixes an exception that is thrown in the rare case that Scarab is set up without the seed data and therefore has no other defined modules. In this case, when a user logs into Scarab, they enter the Global module. However, some code expects the current module to have an owner, and since the Global module did not have one, a NPE was thrown.
The second fix is for the last screen of the issue entry wizard,
Wizard3.vm. Before this change, the reason was considered optional
if a default text attribute was defined for the corresponding issue
type. However, this behaviour is inconsistent with the ViewIssue
screen, which uses the Module.isIssueReasonRequired() method to
determine whether the user must enter a reason when committing
changes to the issue. Now the issue entry screen uses the same
logic.



git-svn-id: http://scarab.tigris.org/svn/scarab@9872 3c7165bb-d2ef-0310-b724-bafd9a86081f