Jorge suggestion to the removed [FIXME]: > How am I supposed to get an activitySet just for the added comment in the > code above? >
You got it as a return value in issue.addComment() :-) so you can use it 
  in your commented out call to NotificationManager.
Anyway... I wonder if this is the correct behaviour.
What about merging the comment creation and the issue-modification if 
they're really the same?
========
AND a bug fix that broke adding a reason in the tab-view screen.


git-svn-id: http://scarab.tigris.org/svn/scarab@9912 3c7165bb-d2ef-0310-b724-bafd9a86081f