SCB1717 fixed:
"ant compile"  (the default target) no more copies any database driver to the WEB-INF/lib directory.
"ant create-db" copies exactly the necessary drivers to the WEB-INF/lib directory.

The driver-template to be used for finding the relevant driver-file(s) is
now named "scarab.database.jdbc.driver.jar" which as in compliance with the
name convention "scarab.database..." for database relevant properties.

Also fixed the documentation and the scarab properties files.

git-svn-id: http://scarab.tigris.org/svn/scarab@10146 3c7165bb-d2ef-0310-b724-bafd9a86081f