fixed SCB1724
Now the lucene search indexes are checked on startup of Scarab. If the
search index directory is empty or contains only the segments file, 
we assume, that the search index has not been created yet, or it has 
previously been deleted on purpose.

In that case the search index is rebuilt in a background thread.

The manual search index rebuild still works as before.


git-svn-id: http://scarab.tigris.org/svn/scarab@10147 3c7165bb-d2ef-0310-b724-bafd9a86081f