[SCB1691] Refactored the GlobalParameterManager for better cache utilizations and code reuse.
Now an empty String is put into the cache, if a parameter has no entry in the database.
Before that, each query for an unset parameter resulted in a db hit because nothing was cached.
At present most parameters are not stored in the db - at least not specific for a module.
So this saves a lot of db-hits.

There are now only two central private methods left which are cached. The public methods are calling this methods.

git-svn-id: http://scarab.tigris.org/svn/scarab@10325 3c7165bb-d2ef-0310-b724-bafd9a86081f