Adds support for multi-part MIME email for sending plain-text-with-HTML email. Modern email clients will render the HTML section, older clients will display the plain text.
For any velocity template used to compose a plain text email, a corresponding template can be used to define an equivalent HTML email. If the HTML template exists and this feature is enabled, a multi-part email is sent with both plain text and HTML sections. If not enabled or if there is no HTML template, only a plain text section is sent. The HTML template is expected to have the same pathname as the plain text template but with "HTML" appended to its basename.

A new HTML template for notifications is included (IssueActivityHtml.vm) so notification email will be multi-part. No other HTML templates have been added so they'll continue to be plain text.

To make this work, I had to get rid of commons-email-1.0-rc5-patched.jar and replace it with the 1.0 release.

While doing that, I made sure that SMTP auth worked (the published patch to 1.0-rc5 didn't). This is now done by setting the authenticator from Scarab so no special commons-email is required.

I also added the facility to enable SMTP debug. This will be very helpful when trouble-shooting problems with sending email.

New properties are: system.email.html and system.mail.smtp.debug



git-svn-id: http://scarab.tigris.org/svn/scarab@10363 3c7165bb-d2ef-0310-b724-bafd9a86081f